Drop die scheme registration
We no longer encourage using dies with clients. Since only a single
type can be registered for a GVK, only the die or the actual type can
be registered with the same scheme. While this can be managed, it isn't
worth the complexity and possible user confusion.

The DeepCopyObject() method can be used to access the resource for the
die, and that can be used any place that interacts with a scheme lookup.
Users who want to use dies directly can manually register the die types.

Signed-off-by: Scott Andrews <scott@andrews.me>