Generate die fields from type ast
Fields are no longer manually defined for dies. The runtime type is
resolved and each exported field has a helper method created, unless
explicitly ignored.

The `+die` marker no longer is defined at the package level, and now
marks the type the die is created for.

For a new struct this looks like:

```go
// +die:object=true
type MyResource struct {
    ...
}
```

And when creating a die for a type defined in a different package:

```go
// +die:object=true
type ExternalResource = pkgalias.ExternalResource
```

Signed-off-by: Scott Andrews <scott@andrews.me>