Add DieStampAt (#105)
Experimental method to stamp the resource at locations defined by a JSON
path expression. The current implementation is crud and requires the
callback function argument type to be compatible with the type
discovered on the resource. The k8s.io/client-go jsonpath implementation
is used.

    var d *diecorev1.PodDie = ...

    d.DieStampAt("$.spec.containers[?(@.name == 'sidecar')]", func(r corev1.Container) {
        r.Env[0].Value = "sidecar"
    })

JSON path enables a user to dive deep into the strucutre of a resource
without interacting with each layer. Expressions can also be filtered to
only interact with a specific subset of the resource.

Signed-off-by: Scott Andrews <scott@andrews.me>