Seal dies to enable diff and patch operations
DieSeal stashes the current value of the die for later use in DieDiff
and DiePatch operations. DieDiff use cmp.Diff to generate a textual diff
of changes made to the die since it was seeled. DiePatch generates patch
bytes to go from the sealed to current value.

Note: diff and patch operations are only meaningful from the root die.
Dies used while stamping do not contain the original sealed value.

Individual patch implementation must be dynamically registered before
use. For the known patch types, this can be done by importing the
respective package. A meaningful error is provided if the type is not
imported. Users may implement additional or alternative patch types.

Signed-off-by: Scott Andrews <scott@andrews.me>