Create owner references for types not in scheme
Duck typed resources are objects that represent many resource types in a
cluster. While they are strongly typed, they are not registered in the
scheme.

When creating a owner reference from an object, if the owner is not in
the scheme, inspect the local TypeMeta. Continue to panic if no value is
available.

Signed-off-by: Scott Andrews <scott@andrews.me>