Add DieFeedDuck/DieReleaseDuck
DieFeedDuck accepts any value and attempts to feed it into the die.
Internally, it marshals the value and feeds the resulting json.
DieReleaseDuck is similar, but take the value from the die and releases
it into the passed value, returning the same value for convenience.

These methods are particularly useful when converting between types with
similar shapes. Values will be lost where the json structures are not
compatible. Any error will result in a panic.

Signed-off-by: Scott Andrews <scott@andrews.me>